comment and fix a few dev script items
